saho-ai-research/requirements.txt
# Core AI/ML
langchain==0.1.0
openai==1.6.1
anthropic==0.8.1
sentence-transformers==2.2.2

# Vector Database
chromadb==0.4.22
# pinecone-client==3.0.0  # Uncomment for production

# API Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.6

# Web Interface
streamlit==1.29.0
# gradio==4.13.0  # Alternative to Streamlit

# Database & Caching
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1

# Data Processing
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
beautifulsoup4==4.12.2
tiktoken==0.5.2
tqdm==4.66.1

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
ruff==0.1.11
pre-commit==3.6.0

# Monitoring (optional for production)
# prometheus-client==0.19.0
# grafana-api==1.0.3